AaharBackend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.7.0
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.25
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.0.5
django-background-tasks==1.2.5
django-celery==3.3.1
django-compat==1.0.15
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-rename-app==0.1.3
django-timezone-field==4.1.2
djangorestframework==3.12.2
djongo==1.3.4
dnspython==1.16.0
gunicorn==20.0.4
idna==2.10
kombu==3.0.37
prompt-toolkit==3.0.18
pymongo==3.11.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.2.0
requests==2.25.1
six==1.15.0
sqlparse==0.2.4
tzlocal==2.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
